+++ src/components/dialogs/index.ts
@@ -3,4 +3,3 @@ export * from "./BottomSheet";
 export { default as BottomSheet } from "./BottomSheet";
 export { default as ConfirmDialog } from "./ConfirmDialog";
 export { default as DeleteDayDialog } from "./DeleteDayDialog";
-export { default as ManageDaySheet } from "./ManageDaySheet";