+++ src/screens/Details/DetailsScreen.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from "react-i18next";
 import { ScrollView, Share, StyleSheet, View } from "react-native";
 import { Menu, Text } from "react-native-paper";
 
-import { type BottomSheetRef, DeleteDayDialog, ManageDaySheet } from "@components/dialogs";
+import { type BottomSheetRef, DeleteDayDialog } from "@components/dialogs";
 import { DayIcon } from "@components/icons";
 import { AppBar, Page, Stack } from "@components/layout";
 import { type AppBarMenuRef } from "@components/layout/AppBarMenu";
@@ -51,7 +51,7 @@ const DetailScreen = () => {
   const menuActionRef = useRef<AppBarMenuRef>(null);
   const manageDayRef = useRef<BottomSheetRef>(null);
 
-  const { onDayDelete, onDayEdit, onDayShare } = useDayActions({
+  const { onDayDelete, onDayShare } = useDayActions({
     onDayEditCallback: () => {
       manageDayRef.current?.close();
     },
@@ -76,11 +76,7 @@ const DetailScreen = () => {
     menuActionRef.current?.close();
     if (!selectedDay) return;
 
-    manageDayRef.current?.open();
-  };
-
-  const onEditCancel = () => {
-    manageDayRef.current?.close();
+    navigation.push("DayFormScreen", { day: selectedDay });
   };
 
   const onSharePress = () => {
@@ -133,14 +129,14 @@ const DetailScreen = () => {
       unitLabel: t("common:timeUnits.day", { count: dateCount.stats.days }),
     },
     {
-      number: dateCount.stats.months,
-      unit: "month",
-      unitLabel: t("common:timeUnits.month", { count: dateCount.stats.months }),
-    },
-    {
       number: dateCount.stats.weeks,
       unit: "week",
       unitLabel: t("common:timeUnits.week", { count: dateCount.stats.weeks }),
+    },
+    {
+      number: dateCount.stats.months,
+      unit: "month",
+      unitLabel: t("common:timeUnits.month", { count: dateCount.stats.months }),
     },
     {
       number: dateCount.stats.years,
@@ -239,12 +235,6 @@ const DetailScreen = () => {
       </ScrollView>
 
       <SwapDetailsTheme>
-        <ManageDaySheet
-          ref={manageDayRef}
-          day={selectedDay}
-          onCancel={onEditCancel}
-          onEdit={onDayEdit}
-        />
         <DeleteDayDialog
           day={deletedDay}
           visible={!!deletedDay}